requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-dotenv==1.0.1
stripe==10.7.0
typer==0.12.5
requests==2.32.3
pydantic==2.9.2
itsdangerous==2.2.0
